rename loading to spp_parser to avoid confusion
